Howard windows updates (#2953)
* update my win conda config

* we need <stdexcept>

* cast away some silly warnings

* ignore conversion warnings in e57 library on windows

* condition e57 ignoration

+++ filters/StatsFilter.hpp
@@ -89,7 +89,7 @@ public:
     {
         if (M2 == 0 || m_cnt <= 2 || !m_advanced)
             return 0.0;
-        double c(m_cnt);
+        double c((double)m_cnt);
         return populationSkewness() * std::sqrt(c) * std::sqrt(c - 1) / (c - 2);
     }
     double skewness() const
@@ -112,14 +112,14 @@ public:
     {
         if (M2 == 0 || m_cnt <= 3 || !m_advanced)
             return 0;
-        double c(m_cnt);
+        double c((double)m_cnt);
         return populationKurtosis() * (c + 1) * (c - 1) / ((c - 2) * (c - 3));
     }
     double sampleExcessKurtosis() const
     {
         if (M2 == 0 || m_cnt <= 3 || !m_advanced)
             return 0;
-        double c(m_cnt);
+        double c((double)m_cnt);
         return sampleKurtosis() - 3 * (c - 1) * (c - 1) / ((c - 2) * (c - 3));
     }
     double kurtosis() const
